# backend/src/main/resources/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Amena Backend API
  description: API specification for the Amena Backend Service
  version: 1.0.0
  contact:
    name: Amena Team
    email: support@amena.com
servers:
  - url: http://localhost:8081
    description: Development server
...
paths:
...
  /api/users:
    get:
      tags:
        - Users
      summary: List all users
      description: Returns a paginated list of users
      operationId: getUsers
      parameters:
        - name: page
          in: query
          description: Page number (0-indexed)
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          description: Page size
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: search
          in: query
          description: Search term
          schema:
            type: string
        - name: role
          in: query
          description: Role filter
          schema:
            type: string
        - name: sortBy
          in: query
          description: Sort field
          schema:
            type: string
            default: createdAt
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            default: desc
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPage"
    post:
      tags:
        - Users
      summary: Create a new user
      description: Creates a new user with the provided details
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User with this email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/users/{userId}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: Returns a single user by their ID
      operationId: getUserById
      parameters:
        - name: userId
          in: path
          required: true
          description: The user's unique identifier
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags:
        - Users
      summary: Update user
      description: Updates an existing user's details
      operationId: updateUser
      parameters:
        - name: userId
          in: path
          required: true
          description: The user's unique identifier
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - Users
      summary: Delete user
      description: Deletes a user by their ID
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          required: true
          description: The user's unique identifier
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: User deleted successfully
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/products:
    get:
      tags:
        - Products
      summary: List all products with pagination and filtering
      operationId: getProducts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: subcategoryId
          in: query
          required: false
          schema:
            type: integer
            format: int64
        - name: sortBy
          in: query
          schema:
            type: string
            default: "createdAt"
        - name: sortOrder
          in: query
          schema:
            type: string
            default: "desc"
        - name: maxStock
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Paginated list of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPage"
    post:
      tags:
        - Products
      summary: Create a new product
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductRequest"
      responses:
        "201":
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
  /api/products/{productId}:
    get:
      tags:
        - Products
      summary: Get product by ID
      operationId: getProductById
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "404":
          description: Product not found
    put:
      tags:
        - Products
      summary: Update product
      operationId: updateProduct
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductRequest"
      responses:
        "200":
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductResponse"
        "404":
          description: Product not found
    delete:
      tags:
        - Products
      summary: Delete product
      operationId: deleteProduct
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Product deleted successfully
        "404":
          description: Product not found
  /api/products/{productId}/images:
    get:
      tags:
        - Products
      summary: List images for a product
      operationId: listProductImages
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: List of product images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductImageResponse"
        "404":
          description: Product not found
  /api/products/{productId}/images/upload:
    post:
      tags:
        - Products
      summary: Upload a product image
      operationId: uploadProductImage
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                primary:
                  type: boolean
                  default: true
      responses:
        "201":
          description: Image uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductImageResponse"
        "404":
          description: Product not found
  /api/products/{productId}/images/{imageId}:
    delete:
      tags:
        - Products
      summary: Delete a product image
      operationId: deleteProductImage
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: imageId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Image deleted successfully
        "404":
          description: Product or image not found
...
  /api/config/{id}:
    get:
      tags:
        - Config
      summary: Get config by ID
      operationId: getConfigById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Config found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigResponse"
        "404":
          description: Config not found
    put:
      tags:
        - Config
      summary: Update config
      operationId: updateConfig
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigRequest"
      responses:
        "200":
          description: Config updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigResponse"
        "201":
          description: Config created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigResponse"
components:
  schemas:
    ConfigRequest:
      type: object
      required:
        - value
      properties:
        value:
          type: string
    ConfigResponse:
      type: object
      required:
        - id
        - value
      properties:
        id:
          type: string
        value:
          type: string
...
    ProductPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/ProductResponse"
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
...
    UserResponse:
      type: object
      required:
        - id
        - email
        - firstName
        - lastName
        - createdAt
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier
        email:
          type: string
          format: email
          description: User's email address
        firstName:
          type: string
          description: User's first name
          minLength: 1
          maxLength: 100
        lastName:
          type: string
          description: User's last name
          minLength: 1
          maxLength: 100
        phone:
          type: string
          description: User's phone number
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
    ProductResponse:
      type: object
      required:
        - id
        - sku
        - name
        - slug
        - price
      properties:
        id:
          type: integer
          format: int64
        sku:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        compareAtPrice:
          type: number
          format: double
        costPrice:
          type: number
          format: double
        categoryId:
          type: integer
          format: int64
        subcategory:
          type: string
        subcategoryIds:
          type: array
          items:
            type: integer
            format: int64
        brand:
          type: string
        weight:
          type: number
          format: double
        isActive:
          type: boolean
        stock:
          type: integer
          format: int32
        isFeatured:
          type: boolean
        metaTitle:
          type: string
        metaDescription:
          type: string
        imageUrl:
          type: string
          description: Public URL of the primary product image
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ProductRequest:
      type: object
      required:
        - sku
        - name
        - slug
        - price
      properties:
        sku:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        compareAtPrice:
          type: number
          format: double
        costPrice:
          type: number
          format: double
        categoryId:
          type: integer
          format: int64
        subcategory:
          type: string
        subcategoryIds:
          type: array
          items:
            type: integer
            format: int64
        brand:
          type: string
        weight:
          type: number
          format: double
        isActive:
          type: boolean
        stock:
          type: integer
          format: int32
        isFeatured:
          type: boolean
        metaTitle:
          type: string
        metaDescription:
          type: string
    CreateUserRequest:
      type: object
      required:
        - email
        - firstName
        - lastName
      properties:
        email:
          type: string
          format: email
          description: User's email address
        firstName:
          type: string
          description: User's first name
          minLength: 1
          maxLength: 100
        lastName:
          type: string
          description: User's last name
          minLength: 1
          maxLength: 100
    UpdateUserRequest:
      type: object
      properties:
        firstName:
          type: string
          description: User's first name
          minLength: 1
          maxLength: 100
        lastName:
          type: string
          description: User's last name
          minLength: 1
          maxLength: 100
        phone:
          type: string
          description: User's phone number
        email:
          type: string
          format: email
          description: User's email address
    UserPage:
      type: object
      required:
        - content
        - totalElements
        - totalPages
        - number
        - size
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/UserResponse"
        totalElements:
          type: integer
          format: int64
          description: Total number of elements
        totalPages:
          type: integer
          description: Total number of pages
        number:
          type: integer
          description: Current page number (0-indexed)
        size:
          type: integer
          description: Page size
        first:
          type: boolean
          description: Whether this is the first page
        last:
          type: boolean
          description: Whether this is the last page
        empty:
          type: boolean
          description: Whether the page is empty
    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
          description: Error occurrence timestamp
        status:
          type: integer
          description: HTTP status code
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Detailed error message
        path:
          type: string
          description: Request path that caused the error
...
    ProductImageResponse:
      type: object
      required:
        - id
        - productId
        - imageUrl
      properties:
        id:
          type: integer
          format: int64
        productId:
          type: integer
          format: int64
        imageUrl:
          type: string
        altText:
          type: string
        sortOrder:
          type: integer
        isPrimary:
          type: boolean
        createdAt:
          type: string
          format: date-time